fix(tests): make the test suite pass on Linux
The suite silently assumed the Windows host it was written on: modules
imported Windows-only APIs at import time, ctypes word widths changed
with the platform, and Windows-form paths were processed through the
host's os.path. 2477 tests now pass on Linux (97 Windows-only tests
skip with reasons); every production change is a no-op on Windows.

- winreg: move imports into the functions that use them (MainWindow,
  msfs_panel_install, ffb_dinput, utils); tap_install takes the hive by
  name; the four dinput tests that monkeypatch winreg skip off-Windows
- simconnect/winstubs: pin DWORD/HRESULT to 4-byte types off-Windows.
  c_ulong is 8 bytes on 64-bit POSIX, which misaligned every
  wire-formatted record - an enumerate-input-events recv decoded as
  RECV_OPEN and reset its own request instead of subscribing
- remove the dead module-level pygetwindow imports (IL2Manager,
  BMSTelemManager) that raised NotImplementedError at import on Linux
- guard the module-level windll setup in winpaths and BMSSharedMem;
  winpaths.get_path raises PathNotFoundException off-Windows, which its
  guarded call sites already handle
- ntpath for Windows-form paths in tap_install (config_link,
  config_label), TapStatusPanel row labels, InstanceSettingsPanel and
  utils.unsafe_install_location_reason: identical to os.path on
  Windows, correct on any host for drive-qualified paths
- skip tests whose subject is inherently Windows-only: ffb_tap (named
  shared memory), the tap write-refusal pair (OSError carries no
  winerror off-Windows), the updater lock tests (a POSIX open handle
  does not block rename/unlink)
- test_tap_workflows: tree() keys normalized to Windows form

bottle (declared in requirements.txt) was simply missing from the
Linux venv; it is installed there, not vendored.

+++ simconnect/winstubs.py
@@ -1,12 +1,15 @@
 # hack to avoid simconnect import error on non-windows platform
 # allows us to run some supporting tools, but not simconnect SDK connection
-from ctypes import c_char, c_ushort, c_ulong, c_void_p, c_char_p
+from ctypes import c_char, c_ushort, c_uint32, c_int32, c_void_p, c_char_p
 
 
-HRESULT = c_ushort
+# widths must match the Windows ABI (ctypes.wintypes) so struct layouts and
+# wire-formatted buffers decode identically on all platforms; c_ulong/c_long
+# are 8 bytes on 64-bit POSIX, which misaligns every 4-byte field
+HRESULT = c_int32
 BYTE = c_char
 WORD = c_ushort
-DWORD = c_ulong
+DWORD = c_uint32
 HANDLE = c_void_p
 LPCSTR = c_char_p
 HWND = c_void_p
